Move command to generate doc to it's own file
